django_project/requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
pika==1.3.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
django-environ==0.11.2